+++ admin/spec/back-end/controllers/test-developer2.spec.js
@@ -5,7 +5,7 @@
 const httpMocks = require('node-mocks-http')
 const sut = require('../../../controllers/test-developer2')
 const checkFormPresenter = require('../../../helpers/check-form-presenter')
-const checkFormV2Service = require('../../../services/check-form-v2.service')
+const checkFormV2Service = require('../../../services/test-developer.service')
 const checkWindowV2Service = require('../../../services/check-window-v2.service')
 
 describe('test developer 2 controller:', () => {
